refactor!: standardize directory structure and namespace (#7, #8)
