+++ CMakeLists.txt
@@ -17,19 +17,22 @@ target_compile_options(compile_settings INTERFACE
 add_library(calsht src/calsht.cpp)
 target_link_libraries(calsht compile_settings)
 target_include_directories(calsht PUBLIC
-  $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/src>
+  $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
   $<INSTALL_INTERFACE:include>)
 target_compile_definitions(calsht INTERFACE
   $<BUILD_INTERFACE:INDEX_FILE_PATH="${PROJECT_BINARY_DIR}">
-  $<INSTALL_INTERFACE:INDEX_FILE_PATH="${CMAKE_INSTALL_PREFIX}/share/data">
+  $<INSTALL_INTERFACE:INDEX_FILE_PATH="${CMAKE_INSTALL_PREFIX}/share/data/mahjong">
   PUBLIC $<$<BOOL:${ENABLE_NYANTEN}>:ENABLE_NYANTEN>)
-set_target_properties(calsht PROPERTIES PUBLIC_HEADER src/calsht.hpp)
 install(TARGETS calsht compile_settings EXPORT Calsht)
+install(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/include/mahjong DESTINATION include)
 install(FILES
   ${PROJECT_BINARY_DIR}/index_s.bin
   ${PROJECT_BINARY_DIR}/index_h.bin
-  DESTINATION share/data)
-install(EXPORT Calsht FILE CalshtConfig.cmake DESTINATION share/cmake)
+  DESTINATION share/data/mahjong)
+install(EXPORT Calsht
+  FILE CalshtConfig.cmake
+  NAMESPACE Mahjong::
+  DESTINATION share/cmake/mahjong)
 
 add_executable(mkind src/mkind.cpp)
 target_compile_definitions(mkind PRIVATE
@@ -37,6 +40,6 @@ target_compile_definitions(mkind PRIVATE
   $<$<BOOL:${TBB_FOUND}>:ENABLE_PARALLEL>)
 target_link_libraries(mkind compile_settings $<$<BOOL:${TBB_FOUND}>:TBB::tbb>)
 
-add_executable(sample src/sample.cpp)
-target_compile_definitions(sample PRIVATE $<$<BOOL:${FIX_RANDOM_SEED}>:FIX_RANDOM_SEED>)
-target_link_libraries(sample calsht)
+add_executable(example example/example.cpp)
+target_compile_definitions(example PRIVATE $<$<BOOL:${FIX_RANDOM_SEED}>:FIX_RANDOM_SEED>)
+target_link_libraries(example calsht)
