+++ README.ja.md
@@ -144,12 +144,12 @@ int main()
 1
 ```
 
-## サンプルプログラム
+## 実行例
 
 手牌をランダムに生成し, シャンテン数ごとの出現率とシャンテン数の期待値を計算します.
 
 ```
-$ ./sample 14 100000000 0
+$ ./example 14 100000000 0
   -1         278    0.000278
    0       69553    0.069553
    1     2334287    2.334287
